chore(db):  Переключение БД на PostgreSQL с in-memory - в TaskExceptionHandler область обработки сужена до TaskController; - модификация application.yml и pom.xml;

+++ src/main/java/ru/t1/ismailov/taskmanager/controller/TaskController.java
@@ -32,18 +32,18 @@ public class TaskController {
     }
 
     @GetMapping("/{id}")
-    public Task getTaskById(@PathVariable Long id) {
+    public Task getTaskById(@PathVariable Integer id) {
         return service.getTaskById(id);
     }
 
     @PutMapping("/{id}")
-    public Task updateTask(@PathVariable Long id, @RequestBody Task task) {
+    public Task updateTask(@PathVariable Integer id, @RequestBody Task task) {
         return service.updateTask(id, task);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTask(@PathVariable Long id) {
+    public void deleteTask(@PathVariable Integer id) {
         service.removeTask(id);
     }
